improve annotation clean up

+++ src/Shared/Presentation/OpenApi/Processor/PathsFeatureProcessor.php
@@ -7,22 +7,20 @@ use App\Shared\Presentation\OpenApi\Routing\Attribute\ApiRouteTrait;
 use App\Shared\Presentation\Publisher\FeaturePublisherContainer;
 use OpenApi\Analysis;
 use OpenApi\Annotations\Operation;
-use OpenApi\Annotations\PathItem;
 use OpenApi\Generator;
 use OpenApi\Processors\ProcessorInterface;
 
 readonly class PathsFeatureProcessor implements ProcessorInterface
 {
+    use ProcessorTrait;
+
     public function __construct(private FeaturePublisherContainer $publishers)
     {
     }
 
     public function __invoke(Analysis $analysis): void
     {
-        /** @var PathItem[] $paths */
-        $pathItems = $analysis->openapi->paths;
-
-        foreach ($pathItems as $index => $pathItem) {
+        foreach ($analysis->openapi->paths as $index => $pathItem) {
             /** @var Operation[]|ApiRouteTrait[] $methods */
             $methods = [
                 $pathItem->post,
@@ -53,8 +51,17 @@ readonly class PathsFeatureProcessor implements ProcessorInterface
                     }
                 } catch (FeatureDisabledError) {
                     $analysis->openapi->paths[$index]->{$method->method} = Generator::UNDEFINED;
-                    $analysis->annotations->detach($pathItem);
+                    $this->detachAnnotationRecursively($method, $analysis);
                 }
+            }
+
+            if (Generator::isDefault($pathItem->post)
+                && Generator::isDefault($pathItem->get)
+                && Generator::isDefault($pathItem->put)
+                && Generator::isDefault($pathItem->patch)
+                && Generator::isDefault($pathItem->delete)) {
+                unset($analysis->openapi->paths[$index]);
+                $analysis->annotations->detach($pathItem);
             }
         }
     }
